Fix dtype mismatch during KD

+++ modules/trainer_teacher.py
@@ -186,7 +186,7 @@ def teacher_adaptive_update(
                 feat_kd_loss = torch.tensor(0.0, device=cfg["device"])
                 if la_mode and cfg.get("feat_kd_alpha", 0) > 0 and not isinstance(mbm, IB_MBM):
                     s_flat = s_feat.view(s_feat.size(0), -1)
-                    f_flat = fsyn.detach().view(fsyn.size(0), -1)
+                    f_flat = fsyn.detach().view(fsyn.size(0), -1).to(s_flat.dtype)
                     if s_flat.size(1) == f_flat.size(1):
                         feat_kd_loss = torch.nn.functional.mse_loss(s_flat, f_flat)
                     else:
